requirements.txt
feedparser==6.0.12
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pyyaml==6.0.1
openai>=1.12.0
schedule==1.2.0
python-dotenv==1.0.0
# Social posting service dependencies
fastapi==0.109.0
uvicorn==0.27.0
requests-oauthlib==1.3.1
pydantic>=2.0.0
